requirements.txt
requests>=2.31.0
anthropic>=0.3.0
pillow>=10.0.0
pytest>=7.4.0  # for testing
